feat(config): add support for config files

+++ phantomrelayd/src/collector/manager.rs
@@ -17,5 +17,5 @@ pub struct HealthyProxy {
     pub ip: String,
     pub port: u16,
     pub metadata: PorxyMetadata,
-    pub latency: u128,
+    pub latency: u64,
 }
